Update code for emptyDir recovery (#608)
When deleting the sts and setting cluster initialized to false for
emptyDir failure scenarios. the osd pod simply tried to reconnect to the
cluster. After the cluster is back up, osd is successfully able to
connect but it doesn't recreate the .kibana index as it is only created
when the pod is started.

Changes:

- Delete OSD deploy so that pods will restart and .kibana index will be
recreated successfully on cluster initialization.
- Also wait for current sts to be deleted.